roles/comfyui/files/comfyui-cli/requirements.txt
click>=8.1,<9
requests>=2.31,<3
pyyaml>=6.0,<7
qdrant-client>=1.9,<2
openai>=1.30,<2
